\encoding{latin1}
\name{FieldSim-internal}
\alias{whichgrid}
\alias{C2D}
\alias{brewer.pal}
\alias{colorRampPalette}
\alias{open3d}
\alias{initialize}
\title{Internal FieldSim Functions}
\description{Internal FieldSim functions.}
\note{These are not to be called by the user (or in some cases are 
just waiting for proper documentation to be written).}
\keyword{internal}
