\name{manifold-class}
\docType{class}
\alias{manifold-class}
\alias{initialize,manifold-method}

\title{Manifold class}
\description{
  The \code{manifold} class is a class of the  \pkg{FieldSim} package.  
}

\section{Slots}{
  \describe{
    \item{\code{name}:}{is the name of the manifold (a character string);}
    \item{\code{atlas}:}{is the mesh (a matrix);}
    \item{\code{gridtype}:}{is the  grid type (a character string) to plotting;}
    \item{\code{distance}:}{is the distance set on the manifold (a function);}
    \item{\code{origin}:}{is the origin fixed on the manifold (a matrix).} 
}}	 

\author{Alexandre Brouste}
\keyword{classes}
