\name{plot-methods}
\docType{class}

\alias{plot-methods}
\alias{plot,manifold-method}
\alias{plot,process-method}

\title{plot methods}
\description{
Specific plot method for process class 
}

\seealso{\code{\link{setProcess}},\code{\link{manifold-class}}.}

\author{Alexandre Brouste}
\keyword{classes}
