\encoding{latin1}
\name{printhyper}
\alias{printhyper}
\title{Print fractional hyperboloïd indexed field simulated on the visual grid}
\usage{printhyper(res)}

\arguments{\item{res}{A result of \code{hypersim} function}}
\description{The function \code{printhyper} yields the visual rgl hyperboloïd
indexed field
simulated by function \code{hypersim}.}

\author{
  Alexandre Brouste (\url{http://subaru.univ-lemans.fr/sciences/statist/pages_persos/Brouste/})
   and
  Sophie Lambert-Lacroix (\url{http://ljk.imag.fr/membres/Sophie.Lambert}).
}

\seealso{\code{\link{hypersim}}.}
\references{A. Brouste, J. Istas and S. Lambert-Lacroix (2009). A. Brouste, J. Istas and S. Lambert-Lacroix (2009). On simulation of manifold indexed fractional Gaussian fields.}