\encoding{latin1}
\name{printsphere}
\alias{printsphere}
\title{Print spherical fractional field simulated on the visual grid}
\usage{printsphere(res)}

\arguments{\item{res}{A result of \code{spheresim} function}}
\description{The function \code{printsphere} yields the visual rgl
sphere indexed field
simulated by function \code{spheresim}.}


\author{
Alexandre Brouste (\url{http://subaru.univ-lemans.fr/sciences/statist/pages_persos/Brouste/})
   and
  Sophie Lambert-Lacroix (\url{http://ljk.imag.fr/membres/Sophie.Lambert}).
}

\seealso{\code{\link{spheresim}}.}
\references{A. Brouste, J. Istas and S. Lambert-Lacroix (2009). A. Brouste, J. Istas and S. Lambert-Lacroix (2009). On simulation of manifold indexed fractional Gaussian fields.}

