\name{show-methods}
\docType{class}

\alias{show-methods}
\alias{print-methods}
\alias{show,manifold-method}
\alias{print,manifold-method}

\title{show methods}
\description{Classical show and print methods available.	 
}

\seealso{\code{\link{process-class}},\code{\link{manifold-class}}.}

\author{Alexandre Brouste}
\keyword{classes}
