\encoding{latin1}
\name{FieldSim-internal}
\alias{gamma2}
\alias{quadvaraux}
\alias{vf}
\alias{sphvf}
\title{Internal FieldSim Functions}
\description{Internal FieldSim functions.}
\note{These are not to be called by the user (or in some cases are 
just waiting for proper documentation to be written).}
\keyword{internal}
